requirements.txt
alembic==1.0.9
aniso8601==6.0.0
Click==7.0
Flask==1.0.2
Flask-Migrate==2.4.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.9
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2==2.8.1
py-bcrypt==0.4
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.3
uWSGI==2.0.18
Werkzeug==0.15.2
